+++ dmr/streaming/__init__.py
@@ -1,4 +1,7 @@
 from dmr.streaming.controller import StreamingController as StreamingController
+from dmr.streaming.endpoint import Streaming as Streaming
+from dmr.streaming.endpoint import modify as modify
+from dmr.streaming.endpoint import validate as validate
 from dmr.streaming.exceptions import StreamingCloseError as StreamingCloseError
 from dmr.streaming.metadata import (
     streaming_response_spec as streaming_response_spec,