+++ tests/test_unit/test_streaming/test_sse/test_sse_event_fields.py
@@ -8,7 +8,7 @@ import pytest
 from dmr.exceptions import ValidationError
 from dmr.plugins.pydantic import PydanticSerializer
 from dmr.serializer import BaseSerializer
-from dmr.streaming import StreamingResponse
+from dmr.streaming import Streaming, StreamingResponse
 from dmr.streaming.sse import SSEController, SSEvent
 from dmr.streaming.sse.validation import check_event_field
 from dmr.test import DMRAsyncRequestFactory
@@ -193,7 +193,7 @@ async def test_wrong_chars_skipped_when_disabled(
     class _ClassBasedSSE(
         SSEController[serializer],  # type: ignore[valid-type]
     ):
-        validate_events = False
+        extras = Streaming(validate_events=False)
 
         async def get(self) -> AsyncIterator[SSEvent[int]]:
             return self._events()
@@ -219,7 +219,7 @@ async def test_custom_event_skipped_when_disabled(
     """Ensures that custom types are not validated when disabled either."""
 
     class _ClassBasedSSE(SSEController[PydanticSerializer]):
-        validate_events = False
+        extras = Streaming(validate_events=False)
 
         async def get(self) -> AsyncIterator[_CustomEvent]:
             return self._events()